style: put some styles in chatting together
스타일이 여러파일 있길래 하나로 합침
