--- front/src/components/GamePlay/CanvasSection/WordChoice/WordSet.style.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const WordSetStyle = styled.div`
-  width: 100%;
-  height: 3rem;
-  display: flex;
-  justify-content: space-between;
-  margin: 20%;
-`;
-
-export default WordSetStyle;
